代理上传不再使用流的 available() 获取文件大小，而是从源头传入实际的大小。避免某些情况 available() 不准确导致上传异常的 bug

+++ src/main/java/im/zhaojun/zfile/core/exception/system/UploadFileFailSystemException.java
@@ -21,13 +21,13 @@ public class UploadFileFailSystemException extends SystemException {
 
     private final String uploadPath;
 
-    private final Integer inputStreamAvailable;
+    private final Long inputStreamAvailable;
 
     private final int responseCode;
 
     private final String responseBody;
 
-    public UploadFileFailSystemException(StorageTypeEnum storageTypeEnum, String uploadPath, Integer inputStreamAvailable, int responseCode, String responseBody) {
+    public UploadFileFailSystemException(StorageTypeEnum storageTypeEnum, String uploadPath, Long inputStreamAvailable, int responseCode, String responseBody) {
         super(ErrorCode.BIZ_UPLOAD_FILE_ERROR);
         this.storageTypeEnum = storageTypeEnum;
         this.uploadPath = uploadPath;
